backend/requirements.txt
arrow==1.3.0
blessed==1.20.0
croniter==2.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
func-timeout==4.3.5
minio==7.2.0
openpyxl==3.1.2
pandas==2.1.4
python-jose==3.3.0
python-keycloak==3.7.0
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.1.0